backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
edge-tts==6.1.8
azure-cognitiveservices-speech==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
websockets==12.0
python-socketio==5.10.0
slowapi==0.1.9
redis==5.0.1
bleach==6.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Logging dependencies
python-json-logger==2.0.7

# Performance monitoring dependencies
psutil==5.9.6
